+++ test/widget/features/companion/roasty_static_test.dart
@@ -1,13 +1,17 @@
+import 'package:brew_path/app/app_theme.dart';
 import 'package:brew_path/features/companion/domain/roasty_state.dart';
 import 'package:brew_path/features/companion/presentation/roasty.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-Widget _wrap(Widget child, {bool disableAnimations = false}) => MediaQuery(
-  data: MediaQueryData(disableAnimations: disableAnimations),
-  child: Directionality(
-    textDirection: TextDirection.ltr,
-    child: Center(child: child),
+Widget _wrap(Widget child, {bool disableAnimations = false}) => Theme(
+  data: AppTheme.cupping,
+  child: MediaQuery(
+    data: MediaQueryData(disableAnimations: disableAnimations),
+    child: Directionality(
+      textDirection: TextDirection.ltr,
+      child: Center(child: child),
+    ),
   ),
 );